+++ storage/src/main/java/org/apache/uniffle/storage/util/ShuffleStorageUtils.java
@@ -206,7 +206,7 @@ public class ShuffleStorageUtils {
 
   // index file header is $PartitionNum | [($PartitionId | $PartitionFileLength | $PartitionDataFileLength), ] | $CRC
   public static long getIndexFileHeaderLen(int partitionNum) {
-    return 4 + (4 + 8 + 8) * partitionNum + 8;
+    return 4 + (4 + 8 + 8) * (long) partitionNum + 8;
   }
 
   public static int getHeaderCrcLen() {
